requirements.txt
pyjwt
gym
numpy
pandas
matplotlib
torch==1.6.0
torchvision==0.7.0